test: refactor error taxonomy integration tests
