+++ VCL/src/test/java/io/velocitycareerlabs/integration/ErrorTaxonomyBackwardCompatibilityBaselineTest.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.velocitycareerlabs.backwardscompatibility
+package io.velocitycareerlabs.integration
 
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
@@ -74,7 +74,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
             val missingDidDeepLink = VCLDeepLink("velocity-network://${entryPoint.schemePath}")
 
             listOf(VCLDeepLink("not a url"), missingDidDeepLink).forEach { deepLink ->
-                val error = getEntryPointError(entryPoint, deepLink)
+                val error = getLegacyEntryPointError(entryPoint, deepLink)
 
                 assertEquals(VCLErrorCode.SdkError.value, error.errorCode)
                 assertTrue(error.message!!.contains("did was not found"))
@@ -88,7 +88,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
             val deepLink = VCLDeepLink(
                 "https://example.com/${entryPoint.schemePath}?${entryPoint.didParam}=did:example:entity"
             )
-            val error = getEntryPointError(entryPoint, deepLink)
+            val error = getLegacyEntryPointError(entryPoint, deepLink)
 
             assertEquals(VCLErrorCode.SdkError.value, error.errorCode)
             assertEquals(entryPoint.endpointNullMessage, error.message)
@@ -116,7 +116,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
             val deepLink = VCLDeepLink(
                 "velocity-network://${entryPoint.schemePath}?${entryPoint.didParam}=did:example:entity"
             )
-            val error = getEntryPointError(entryPoint, deepLink)
+            val error = getLegacyEntryPointError(entryPoint, deepLink)
 
             assertEquals(VCLErrorCode.SdkError.value, error.errorCode)
             assertEquals(entryPoint.endpointNullMessage, error.message)
@@ -125,7 +125,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
 
     @Test
     fun missingDirectRequestDidPreservesLegacyDidMessage() {
-        val error = getCredentialManifestDescriptorError(
+        val error = getLegacyCredentialManifestDescriptorError(
             descriptor = credentialManifestDescriptorByService(did = ""),
         )
 
@@ -145,8 +145,8 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
                     "request_uri=ftp%3A%2F%2Fexample.com%2Frequest" +
                     "&${entryPoint.didParam}=did:example:entity"
             )
-            val malformedRequestUri = getEntryPointError(entryPoint, malformedRequestUriDeepLink)
-            val disallowedSchemeRequestUri = getEntryPointError(entryPoint, disallowedSchemeDeepLink)
+            val malformedRequestUri = getLegacyEntryPointError(entryPoint, malformedRequestUriDeepLink)
+            val disallowedSchemeRequestUri = getLegacyEntryPointError(entryPoint, disallowedSchemeDeepLink)
 
             assertEquals(VCLErrorCode.SdkError.value, malformedRequestUri.errorCode)
             assertTrue(malformedRequestUri.message!!.contains("no protocol: not-a-url"))
@@ -160,7 +160,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun transportFailureReturnsSdkErrorWithNetworkStatusOnly() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(
                     requestFailure = UnknownHostException("offline"),
@@ -177,7 +177,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     fun requestEndpoint401And403PreserveHttpStatusAndPayloadErrorCode() {
         entryPoints.forEach { entryPoint ->
             listOf(401, 403).forEach { statusCode ->
-                val error = getEntryPointError(
+                val error = getLegacyEntryPointError(
                     entryPoint,
                     router = defaultRouter(entryPoint).copy(
                         requestStatusCode = statusCode,
@@ -202,7 +202,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
 
         entryPoints.forEach { entryPoint ->
             listOf(400, 404, 409, 410, 422, 500, 502).forEach { statusCode ->
-                val error = getEntryPointError(
+                val error = getLegacyEntryPointError(
                     entryPoint,
                     router = defaultRouter(entryPoint).copy(
                         requestStatusCode = statusCode,
@@ -220,7 +220,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun plainTextRequestEndpointRejectionsDefaultToSdkErrorWithHttpStatusAndPayloadMessage() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(
                     requestStatusCode = 500,
@@ -243,7 +243,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
         }.toString()
 
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(
                     requestStatusCode = 422,
@@ -262,7 +262,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun emptyRequestEndpointResponseReturnsSdkError() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(requestPayload = ""),
             )
@@ -274,7 +274,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun malformedRequestEndpointResponseReturnsSdkError() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(requestPayload = "not json"),
             )
@@ -286,7 +286,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun missingExpectedRequestFieldsReturnSdkErrorAfterEmptyJwtIsDecoded() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(requestPayload = "{}"),
             )
@@ -300,7 +300,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun didResolutionNetworkFailurePropagatesSdkErrorAndStatusFromNetwork() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(
                     didDocumentStatusCode = 404,
@@ -318,7 +318,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun invalidDidDocumentShapeReturnsSdkErrorAtRequestValidation() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(didDocumentPayload = "not json"),
             )
@@ -331,7 +331,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun missingDidDocumentVerificationMaterialReturnsSdkError() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(didDocumentPayload = "{}"),
             )
@@ -346,7 +346,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun verifiedProfileLookupFailurePropagatesNetworkErrorDetails() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(
                     verifiedProfileStatusCode = 404,
@@ -366,7 +366,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     @Test
     fun emptyVerifiedProfileFailsServiceTypeVerification() {
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(verifiedProfilePayload = "{}"),
             )
@@ -384,7 +384,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
                 EntryPoint.Issuing -> VerifiedProfileMocks.VerifiedProfileInspectorJsonStr
                 EntryPoint.Presentation -> VerifiedProfileMocks.VerifiedProfileIssuerJsonStr1
             }
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 router = defaultRouter(entryPoint).copy(verifiedProfilePayload = wrongServiceProfile),
             )
@@ -401,7 +401,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
     fun duplicateQueryParamsUseLastDidValueAtSdkEntryPoint() {
         entryPoints.forEach { entryPoint ->
             val router = defaultRouter(entryPoint)
-            val vcl = initializedVcl(router)
+            val vcl = initializedVcl(router, errorCodeCompatibilityMode = VCLErrorCodeCompatibilityMode.Legacy)
             val deepLink = VCLDeepLink(
                 "velocity-network://${entryPoint.schemePath}?request_uri=${entryPoint.encodedRequestUri}" +
                     "&${entryPoint.didParam}=did:example:first&${entryPoint.didParam}=${entryPoint.lastDid}"
@@ -429,7 +429,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
                 "velocity-network://${entryPoint.schemePath}?request_uri=${entryPoint.encodedRequestUri}" +
                     "&${entryPoint.didParam}=not-a-did"
             )
-            val error = getEntryPointError(entryPoint, deepLink)
+            val error = getLegacyEntryPointError(entryPoint, deepLink)
 
             assertEquals(entryPoint.mismatchErrorCode, error.errorCode)
         }
@@ -442,7 +442,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
                 "velocity-network://${entryPoint.schemePath}?request_uri=${entryPoint.encodedRequestUri}" +
                     "&${entryPoint.didParam}=did:example:wrong"
             )
-            val error = getEntryPointError(entryPoint, deepLink)
+            val error = getLegacyEntryPointError(entryPoint, deepLink)
 
             assertEquals(entryPoint.mismatchErrorCode, error.errorCode)
         }
@@ -456,7 +456,7 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
         )
 
         entryPoints.forEach { entryPoint ->
-            val error = getEntryPointError(
+            val error = getLegacyEntryPointError(
                 entryPoint,
                 jwtVerificationResult = VCLResult.Failure(expectedError),
             )
@@ -466,318 +466,4 @@ internal class ErrorTaxonomyBackwardCompatibilityBaselineTest {
         }
     }
 
-    private enum class EntryPoint {
-        Issuing,
-        Presentation,
-    }
-
-    private val entryPoints = listOf(EntryPoint.Issuing, EntryPoint.Presentation)
-
-    private val EntryPoint.defaultDeepLink: VCLDeepLink
-        get() = when (this) {
-            EntryPoint.Issuing -> DeepLinkMocks.CredentialManifestDeepLinkDevNet
-            EntryPoint.Presentation -> DeepLinkMocks.PresentationRequestDeepLinkDevNet
-        }
-
-    private val EntryPoint.endpointNullMessage: String
-        get() = when (this) {
-            EntryPoint.Issuing -> "credentialManifestDescriptor.endpoint = null"
-            EntryPoint.Presentation -> "presentationRequestDescriptor.endpoint = null"
-        }
-
-    private val EntryPoint.mismatchErrorCode: String
-        get() = when (this) {
-            EntryPoint.Issuing -> VCLErrorCode.MismatchedRequestIssuerDid.value
-            EntryPoint.Presentation -> VCLErrorCode.MismatchedPresentationRequestInspectorDid.value
-        }
-
-    private val EntryPoint.didParam: String
-        get() = when (this) {
-            EntryPoint.Issuing -> "issuerDid"
-            EntryPoint.Presentation -> "inspectorDid"
-        }
-
-    private val EntryPoint.schemePath: String
-        get() = when (this) {
-            EntryPoint.Issuing -> "issue"
-            EntryPoint.Presentation -> "inspect"
-        }
-
-    private val EntryPoint.encodedRequestUri: String
-        get() = when (this) {
-            EntryPoint.Issuing -> DeepLinkMocks.CredentialManifestRequestUriStr
-            EntryPoint.Presentation -> DeepLinkMocks.PresentationRequestRequestUriStr
-        }
-
-    private val EntryPoint.lastDid: String get() = "did:example:last"
-
-    private fun defaultRouter(entryPoint: EntryPoint): BaselineHttpRouter =
-        when (entryPoint) {
-            EntryPoint.Issuing -> BaselineHttpRouter()
-            EntryPoint.Presentation -> BaselineHttpRouter(
-                verifiedProfilePayload = VerifiedProfileMocks.VerifiedProfileInspectorJsonStr,
-                requestPayload = PresentationRequestMocks.EncodedPresentationRequestResponse,
-            )
-        }
-
-    private fun getEntryPointError(
-        entryPoint: EntryPoint,
-        deepLink: VCLDeepLink = entryPoint.defaultDeepLink,
-        router: BaselineHttpRouter = defaultRouter(entryPoint),
-        jwtVerificationResult: VCLResult<Boolean> = VCLResult.Success(true),
-    ): VCLError =
-        when (entryPoint) {
-            EntryPoint.Issuing -> getCredentialManifestError(deepLink, router, jwtVerificationResult)
-            EntryPoint.Presentation -> getPresentationRequestError(deepLink, router, jwtVerificationResult)
-        }
-
-    private fun getCredentialManifestError(
-        deepLink: VCLDeepLink,
-        router: BaselineHttpRouter = BaselineHttpRouter(),
-        jwtVerificationResult: VCLResult<Boolean> = VCLResult.Success(true),
-    ): VCLError {
-        val vcl = initializedVcl(router, jwtVerificationResult)
-        return awaitCredentialManifestError(vcl, credentialManifestDescriptor(deepLink))
-    }
-
-    private fun getCredentialManifestDescriptorError(
-        descriptor: VCLCredentialManifestDescriptor,
-        router: BaselineHttpRouter = BaselineHttpRouter(),
-        jwtVerificationResult: VCLResult<Boolean> = VCLResult.Success(true),
-    ): VCLError {
-        val vcl = initializedVcl(router, jwtVerificationResult)
-        return awaitCredentialManifestError(vcl, descriptor)
-    }
-
-    private fun getPresentationRequestError(
-        deepLink: VCLDeepLink,
-        router: BaselineHttpRouter = BaselineHttpRouter(
-            verifiedProfilePayload = VerifiedProfileMocks.VerifiedProfileInspectorJsonStr,
-            requestPayload = PresentationRequestMocks.EncodedPresentationRequestResponse,
-        ),
-        jwtVerificationResult: VCLResult<Boolean> = VCLResult.Success(true),
-    ): VCLError {
-        val vcl = initializedVcl(router, jwtVerificationResult)
-        return awaitPresentationRequestError(vcl, presentationDescriptor(deepLink))
-    }
-
-    private fun initializedVcl(
-        router: BaselineHttpRouter,
-        jwtVerificationResult: VCLResult<Boolean> = VCLResult.Success(true),
-    ): VCLImpl {
-        val vcl = VCLImpl(router::connectionFor)
-        var initError: VCLError? = null
-        val latch = CountDownLatch(1)
-        vcl.initialize(
-            context = ApplicationProvider.getApplicationContext(),
-            initializationDescriptor = VCLInitializationDescriptor(
-                cacheSequence = cacheSequence.incrementAndGet(),
-                errorCodeCompatibilityMode = VCLErrorCodeCompatibilityMode.Legacy,
-                cryptoServicesDescriptor = VCLCryptoServicesDescriptor(
-                    cryptoServiceType = VCLCryptoServiceType.Injected,
-                    injectedCryptoServicesDescriptor = VCLInjectedCryptoServicesDescriptor(
-                        keyService = VCLKeyServiceMock(),
-                        jwtSignService = VCLJwtSignServiceMock(),
-                        jwtVerifyService = FixedJwtVerifyService(jwtVerificationResult),
-                    ),
-                ),
-            ),
-            successHandler = {
-                latch.countDown()
-            },
-            errorHandler = {
-                initError = it
-                latch.countDown()
-            },
-        )
-        drainMainThreadUntil(latch)
-        assertNull("VCL initialization failed: ${initError?.toJsonObject()}", initError)
-        return vcl
-    }
-
-    private fun awaitCredentialManifestError(
-        vcl: VCLImpl,
-        descriptor: VCLCredentialManifestDescriptor,
-    ): VCLError {
-        var result: VCLError? = null
-        val latch = CountDownLatch(1)
-        vcl.getCredentialManifest(
-            credentialManifestDescriptor = descriptor,
-            successHandler = { manifest ->
-                fail("Credential manifest failure expected: $manifest")
-            },
-            errorHandler = {
-                result = it
-                latch.countDown()
-            },
-        )
-        drainMainThreadUntil(latch)
-        return result ?: error("getCredentialManifest did not invoke errorHandler")
-    }
-
-    private fun awaitPresentationRequestError(
-        vcl: VCLImpl,
-        descriptor: VCLPresentationRequestDescriptor,
-    ): VCLError {
-        var result: VCLError? = null
-        val latch = CountDownLatch(1)
-        vcl.getPresentationRequest(
-            presentationRequestDescriptor = descriptor,
-            successHandler = { presentationRequest ->
-                fail("Presentation request failure expected: $presentationRequest")
-            },
-            errorHandler = {
-                result = it
-                latch.countDown()
-            },
-        )
-        drainMainThreadUntil(latch)
-        return result ?: error("getPresentationRequest did not invoke errorHandler")
-    }
-
-    private fun drainMainThreadUntil(latch: CountDownLatch) {
-        val deadline = System.currentTimeMillis() + 5_000
-        while (latch.count > 0 && System.currentTimeMillis() < deadline) {
-            shadowOf(android.os.Looper.getMainLooper()).idle()
-            latch.await(20, TimeUnit.MILLISECONDS)
-        }
-        shadowOf(android.os.Looper.getMainLooper()).idle()
-        assertEquals("Timed out waiting for callback", 0, latch.count)
-    }
-
-    private fun credentialManifestDescriptor(deepLink: VCLDeepLink) =
-        VCLCredentialManifestDescriptorByDeepLink(
-            deepLink = deepLink,
-            didJwk = DidJwkMocks.DidJwk,
-        )
-
-    private fun credentialManifestDescriptorByService(
-        did: String = DeepLinkMocks.IssuerDid,
-    ) =
-        VCLCredentialManifestDescriptorByService(
-            service = VCLService(
-                JSONObject()
-                    .put(VCLService.KeyId, "${DeepLinkMocks.IssuerDid}#credential-agent-issuer-1")
-                    .put(VCLService.KeyType, "VelocityCredentialAgentIssuer_v1.0")
-                    .put(VCLService.KeyServiceEndpoint, DeepLinkMocks.CredentialManifestRequestDecodedUriStr)
-            ),
-            issuingType = VCLIssuingType.Career,
-            didJwk = DidJwkMocks.DidJwk,
-            did = did,
-        )
-
-    private fun presentationDescriptor(deepLink: VCLDeepLink) =
-        VCLPresentationRequestDescriptor(
-            deepLink = deepLink,
-            didJwk = DidJwkMocks.DidJwk,
-        )
-
-    private class FixedJwtVerifyService(
-        private val result: VCLResult<Boolean>,
-    ) : VCLJwtVerifyService {
-        override fun verify(
-            jwt: VCLJwt,
-            publicJwk: VCLPublicJwk,
-            remoteCryptoServicesToken: VCLToken?,
-            completionBlock: (VCLResult<Boolean>) -> Unit,
-        ) {
-            completionBlock(result)
-        }
-    }
-
-    private data class BaselineHttpRouter(
-        private val verifiedProfilePayload: String = VerifiedProfileMocks.VerifiedProfileIssuerJsonStr1,
-        private val verifiedProfileStatusCode: Int = 200,
-        private val verifiedProfileContentType: String? = Request.ContentTypeApplicationJson,
-        private val requestPayload: String = CredentialManifestMocks.CredentialManifest1,
-        private val requestStatusCode: Int = 200,
-        private val requestContentType: String? = Request.ContentTypeApplicationJson,
-        private val requestFailure: Exception? = null,
-        private val didDocumentPayload: String = DidDocumentMocks.DidDocumentMockStr,
-        private val didDocumentStatusCode: Int = 200,
-        private val didDocumentContentType: String? = Request.ContentTypeApplicationJson,
-    ) {
-        val requestedEndpoints: MutableList<String> = Collections.synchronizedList(mutableListOf())
-
-        fun connectionFor(request: Request): HttpURLConnection {
-            requestedEndpoints.add(request.endpoint)
-            if (isSdkRequestEndpoint(request.endpoint)) {
-                requestFailure?.let { throw it }
-                if (request.endpoint.startsWith("not-a-url")) {
-                    throw MalformedURLException("no protocol: ${request.endpoint}")
-                }
-                if (request.endpoint.startsWith("ftp://")) {
-                    throw MalformedURLException("unknown protocol: ftp")
-                }
-            }
-            val route = routeFor(request.endpoint)
-            return FakeHttpURLConnection(
-                url = URL(request.endpoint),
-                responseCodeValue = route.statusCode,
-                contentTypeValue = route.contentType,
-                payload = route.payload,
-            )
-        }
-
-        private fun routeFor(endpoint: String): Route =
-            when {
-                endpoint.contains("/reference/countries") ->
-                    Route(200, Request.ContentTypeApplicationJson, CountriesMocks.CountriesJson)
-                endpoint.contains("/api/v0.6/credential-types") ->
-                    Route(200, Request.ContentTypeApplicationJson, CredentialTypesMocks.CredentialTypesJson)
-                endpoint.contains("/schemas/") ->
-                    Route(200, Request.ContentTypeApplicationJson, CredentialTypeSchemaMocks.CredentialTypeSchemaJson)
-                endpoint.contains("/verified-profile") ->
-                    Route(verifiedProfileStatusCode, verifiedProfileContentType, verifiedProfilePayload)
-                endpoint.contains("/resolve-did/") ->
-                    Route(didDocumentStatusCode, didDocumentContentType, didDocumentPayload)
-                isSdkRequestEndpoint(endpoint) ->
-                    Route(requestStatusCode, requestContentType, requestPayload)
-                else ->
-                    error("Unhandled HTTP endpoint in baseline test: $endpoint")
-            }
-
-        private fun isSdkRequestEndpoint(endpoint: String): Boolean =
-            endpoint.contains("get-credential-manifest") ||
-                endpoint.contains("get-presentation-request") ||
-                endpoint.startsWith("not-a-url") ||
-                endpoint.startsWith("ftp://")
-    }
-
-    private data class Route(
-        val statusCode: Int,
-        val contentType: String?,
-        val payload: String,
-    )
-
-    private class FakeHttpURLConnection(
-        url: URL,
-        private val responseCodeValue: Int,
-        private val contentTypeValue: String?,
-        private val payload: String,
-    ) : HttpURLConnection(url) {
-        override fun connect() = Unit
-
-        override fun disconnect() = Unit
-
-        override fun usingProxy() = false
-
-        override fun getResponseCode(): Int = responseCodeValue
-
-        override fun getContentType(): String? = contentTypeValue
-
-        override fun getErrorStream(): InputStream? =
-            if (responseCodeValue >= HTTP_OK && responseCodeValue <= 299) {
-                null
-            } else {
-                ByteArrayInputStream(payload.toByteArray())
-            }
-
-        override fun getInputStream(): InputStream =
-            ByteArrayInputStream(payload.toByteArray())
-    }
-
-    private companion object {
-        val cacheSequence = AtomicInteger(1000)
-    }
 }
